refactor line graph into its own component
