+++ client/src/components/InsightsLineGraph.tsx
@@ -1,0 +1,67 @@
+import React from "react";
+import { Line } from "react-chartjs-2";
+import {
+  Chart as ChartJS,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend,
+} from "chart.js";
+import formatLineGraphData from "../utils/formatLineGraphData";
+
+ChartJS.register(
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend
+);
+
+const InsightsLineGraph = ({ insightsData }: { insightsData: any }) => {
+  const formattedData = formatLineGraphData(insightsData);
+
+  const labels = formattedData.months;
+
+  const data = {
+    labels,
+    datasets: [
+      {
+        label: "Applications made",
+        data: formattedData.totalCount,
+        borderColor: "rgba(255, 99, 132)",
+        backgroundColor: "rgba(255, 99, 132, 0.5)",
+      },
+      {
+        label: "Accepted",
+        data: formattedData.acceptedCount,
+        borderColor: "rgb(53, 162, 235)",
+        backgroundColor: "rgba(53, 162, 235, 0.5)",
+      },
+    ],
+  };
+  const options = {
+    tension: 0.5,
+    responsive: true,
+    plugins: {
+      legend: {
+        position: "top" as const,
+      },
+      // title: {
+      //   display: true,
+      //   text: "Monthly application totals",
+      // },
+    },
+  };
+  return (
+    <div className="p-4 border border-borderprimary rounded-md">
+      <Line options={options} data={data} />
+    </div>
+  );
+};
+
+export default InsightsLineGraph;
