+++ client/src/pages/AppsInsightPage.tsx
@@ -1,74 +1,16 @@
 import axios from "axios";
 import React from "react";
 import { Params, useLoaderData } from "react-router-dom";
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend,
-} from "chart.js";
-import { Line } from "react-chartjs-2";
-import formatLineGraphData from "../utils/formatLineGraphData";
 import InsightsTotalCounts from "../components/InsightsTotalCounts";
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend
-);
+import InsightsLineGraph from "../components/InsightsLineGraph";
 
 export const AppsInsightPage = () => {
   const { insightsData } = useLoaderData() as { insightsData: any };
 
-  const formattedData = formatLineGraphData(insightsData);
-
-  const options = {
-    tension: 0.5,
-    responsive: true,
-    plugins: {
-      legend: {
-        position: "top" as const,
-      },
-      // title: {
-      //   display: true,
-      //   text: "Monthly application totals",
-      // },
-    },
-  };
-
-  const labels = formattedData.months;
-
-  const data = {
-    labels,
-    datasets: [
-      {
-        label: "Applications made",
-        data: formattedData.totalCount,
-        borderColor: "rgba(255, 99, 132)",
-        backgroundColor: "rgba(255, 99, 132, 0.5)",
-      },
-      {
-        label: "Accepted",
-        data: formattedData.acceptedCount,
-        borderColor: "rgb(53, 162, 235)",
-        backgroundColor: "rgba(53, 162, 235, 0.5)",
-      },
-    ],
-  };
   return (
     <div>
       <InsightsTotalCounts data={insightsData} />
-      <div className="p-4 border border-borderprimary rounded-md">
-        <Line options={options} data={data} />
-      </div>
+      <InsightsLineGraph insightsData={insightsData} />
     </div>
   );
 };
